updated documentation and small renaming in the package
