+++ users/__init__.py
@@ -4,7 +4,7 @@ This module contains the implementation of the Users package,
 which provides functionality for managing users and rate limiting.
 """
 
-from .ratelimits import RateLimiter
+from .ratelimiter import RateLimiter
 from .users import Users
 from .storage import Storage
 from .constants import USERS_VAULT_CONFIG_PATH, USER_STATUS_ALLOW, USER_STATUS_DENY
