+++ users/users.py
@@ -6,7 +6,7 @@ import json
 from logger import log
 from vault import VaultClient
 from .constants import USERS_VAULT_CONFIG_PATH, USER_STATUS_ALLOW, USER_STATUS_DENY
-from .ratelimits import RateLimiter
+from .ratelimiter import RateLimiter
 from .storage import Storage
 from .exceptions import VaultInstanceNotSet
